Experiment34

+++ triton_kernel/Hierarchical_Attention_Triton.py
@@ -1483,9 +1483,9 @@ def build_tree_topology(seq_len, is_causal=True, device="cuda", dtype=torch.int3
     valid_nodes = all_nodes[should_gather]
     valid_siblings = siblings[should_gather]
 
-    gather_info[valid_nodes, 0] = subtree_ranges[valid_siblings, 0].to(torch.int64)
-    gather_info[valid_nodes, 1] = subtree_ranges[valid_siblings, 1].to(torch.int64)
-    gather_info[valid_nodes, 2] = node_levels[valid_nodes].to(torch.int64) # Now uses the corrected levels
+    gather_info[valid_nodes, 0] = subtree_ranges[valid_siblings, 0]
+    gather_info[valid_nodes, 1] = subtree_ranges[valid_siblings, 1]
+    gather_info[valid_nodes, 2] = node_levels[valid_nodes]
 
 
     ## ===============================================================
